#35 refactor: 차트 데이터 로직 커스텀 훅(useChartPagination)으로 분리
